Fixing some issues with downloads

+++ Emitron/Emitron/Downloads/DownloadService.swift
@@ -178,10 +178,15 @@ extension DownloadService: DownloadAction {
   
   func cancelDownload(contentId: Int) throws {
     do {
+      // 0. If there are some children, then let's cancel all of them first.
+      if let children = try persistenceStore.childContentsForDownloadedContent(with: contentId) {
+        try children.contents.forEach { try cancelDownload(contentId: $0.id) }
+      }
+      
       // 1. Find the download.
       guard let download = try persistenceStore.download(forContentId: contentId) else { return }
       // 2. Is it already downloading?
-      if [.inProgress, .paused].contains(download.state) {
+      if [.inProgress, .paused].contains(download.state) && download.remoteUrl != nil {
         // It's in the download process, so let's ask it to cancel it. The delegate callback will handle deleting the value in the persistence store.
         try downloadProcessor.cancelDownload(download)
       } else {
@@ -198,10 +203,17 @@ extension DownloadService: DownloadAction {
   
   func deleteDownload(contentId: Int) throws {
     do {
+      // 0. If there are some children, the let's delete all of them first.
+      if let children = try persistenceStore.childContentsForDownloadedContent(with: contentId) {
+        try children.contents.forEach { try deleteDownload(contentId: $0.id) }
+      }
+      
       // 1. Find the download
       guard let download = try persistenceStore.download(forContentId: contentId) else { return }
       // 2. Delete the file from disk
-      try deleteFile(for: download)
+      if [.complete].contains(download.state) && download.remoteUrl != nil {
+        try deleteFile(for: download)
+      }
       // 3. Delete the persisted record
       if try !persistenceStore.deleteDownload(withId: download.id) {
         Failure
